Make `fatrpm` generate backwards-compatible RPMs
Right now, the generated rpm digest and payloads depend of the settings
of the Linux distribution being used, which means that on any recent
CentOS we end up creating packages that use the following:

    * File digest: SHA384
    * Payload: xz

A non-gzip payload breaks backwards compatibility, so what this patch
does is making both settings use the most basic (and default) build
settings. Now the generated package is:

    * File digest: MD5
    * Payload: gzip

This should make the rpm installable in any distribution that does rpm.