Don't copy into {} before rename-keys or deep-merge
I put those `into {}` to work around a failure back when we used to
convert the state into a java.map.HashMap before setting it as the
new global state.

This can still be reproduced by just calling rename-keys in the original
state (a java.util.HashMap instance). We get this:

    java.lang.ClassCastException: java.util.HashMap cannot be cast to clojure.lang.IPersistentMap

This patch adds a test to the scenario and changes the original instance
of `EXPR_GLOBAL_STATE_RO` to be a `clojure.lang.PersistentArrayMap`.